arm64: dts: exynos: gs101-pixel: add all S2MPG1x regulators
Most rails are the same between Pixel 6 and Pro, with the following
differences:
    * only Pro has UWB
    * Pro uses l2m, not l14m, for TCXO
    * Pro uses bucka, not l31m, for NFC

Signed-off-by: André Draszik <andre.draszik@linaro.org>

---
Please note that this patch depends on the bindings updates for S2MPG1x
from
https://lore.kernel.org/all/20251110-s2mpg1x-regulators-v4-0-94c9e726d4ba@linaro.org/

I've left out conversion of placeholder regulators for now, because
some more testing is required for such changes.

With this in place (and the corresponding driver updates), the most
noteable change is that the display is turned off at some point during
boot (since we don't have a display driver yet). To avoid that, one
could boot with regulator_ignore_unused or vci_disp (ldo22m) could be
marked as regulator-always-on. I've not done that to ensure power
consumption is reduced by default and the phone stays cool and the
battery charges using the default settings configured by the bootloader.

v2:
* add comments for board-specific and unused rails (Krzysztof)